cmd-podman-build: stop passing the git yum repos
This goes together with Jonathan's commit in
https://github.com/openshift/os/pull/1498 titled "build-node-image.sh:
include yum repos from git as part of build".

There's no point in adding the repo files already in the git repo into
the secret; the node and extension builds already have access to those.
So we'd just be duplicating them.

Another way to think of this is that the secret injection we do here
should be considered equivalent to `get-ocp-repo.sh` in OpenShift CI in
terms of what repos get injected.