cosalib: add Builds().get_build_image_json(); drop use of flatten_image_yaml()
As part of c61da89 we started to allow templating in the image.yaml
file. This means we now need to pass the dict of things to template
into flatten-image-yaml and that is now causing us issues if there
are things to template but not template dict passed in.

One case we were doing this was in our AWS upload where we pick
up some values from image.yaml. Except, we shouldn't have been
picking them up from image.yaml in the src/config as this stage
but rather the image.json that had been baked into the build.

Let's switch over cosalib/aws to pick the info up from image.json
instead. As part of this add a helper the the cosalib Builds library
to make this easier for us in various places to do it. This addition
was conveniently stolen from cmd-compress that already had code
to do this.