mantle/platform/aws: expand API to support windows li images
Expand the API to support building Windows License Included images.
The added methods include creating, attaching, detaching, and deleting
volumes, stopping instances, and creating images from instances.

Update RemoveImage to only delete the s3 object if the bucket and path
are specified to allow the method to be used to delete the WinLI image,
does not have an uploaded s3 object associated with it.

A ReplaceRootVolume method is also added to swap the root volume of an
ec2 instance with a target volume.