Add cosa imageupload-<platform> subcommands
Previously `cosa buildextend-aws --upload` would both build AND
upload to AWS EC2. Now that we are building using OSbuild it
makes more sense to split these stages a bit so we have
`cosa osbuild aws` and then `cosa imageupload-aws` for the uploading
part.

`cosa buildextend-aws` is now just a symlink to `cosa osbuild aws`
so we need these new `imageupload-<platform>` subcommands for the
uploading part.