mantle/platform/aws: update name of describeImage
change `describeImage` to `DescribeImage` to match the general naming
convention for methods in this API. For the most part, methods begin
with a capital letter, which allows them to be exported out of their
packages, but this rule isn't exclusively followed. There are other
methods that could be updated as well, if that's something we want
to do. For now, let's just update this one because it was bugging me.