mantle/api: move random name function from azure to util/common
Move the `randomName` function to the util package so it can be used
elsewhere. Specifically, we'll use it when creating resources along with
the 'aws-winli' AMI, but now it can be used anwhere. The name of the
function is also capitalized so it can be exported outside of the util
package.

Also remove the plog variable. It's not actually used in this package
so linting in CI is throwing errors. We can add it back if we have a
need for it in the future.