Makefile: Don't always build kolet for all architectures
The reason we build kolet for other architectures is so that we
can mix architectures with test initiators and test runners. i.e.
i.e. x86_64 kola running in the pipeline testing against an aarch64
EC2 instance.

This use case isn't too common so let's just not build kolet for
all the different architectures on every architecture. We'll special
case for x86_64 here and let it build for them all, but for the rest
we just build for the architecture that the container build is for.