kola/spawn: support GCP from stream metadata too
The same data is there for GCP as it is for AWS so we might as well
support pulling the info from stream metadata too.

```
cosa kola spawn -b fcos --stream testing -p gcp --gcp-json-key=key
```

GCP is a bit nicer than AWS because when we spawn AWS nodes, and then
shut them down later, the code will try to grab the serial console logs
and for AWS that takes some time (if it ever actually is successful).
For GCP we don't really have that problem, so spawning (and destroying)
a GCP instance is much faster than AWS.