cmd-coreos-prune: handle expired or deleted tags in skopeo inspect
When a tag is expired or been garbage collected in Quay, skopeo inspect
returns a "unknown" error with a message like:

"Tag <tag> was deleted or has expired. To pull, revive via time machine"

Check for this condition and skip raising the exception, instead just
return since the tag has already been GC'd.