mantle/kola: add InstanceType to PlatformOptions for external tests
Add an `InstanceType` field  to `PlatformOptions` to allow external
tests to override the instance type used in `kola run`. This is useful
for cases where a specific test needs to run on a different (potentially
more expensive) instance type. Support for this is currently limited
to the Azure Platform.

Also, fix the `MultiPathDisk` check for the qemu-iso platform. The check
is now correctly performed in the `NewMachineWithOptions` function, since
`MultiPathDisk` is part of `platform.MachineOptions`.