cmd-remote-build-container: add --add-openshift-build-labels
Currently, `oc` looks for non-standard labels for git information, which
it then places in e.g. `image-references` in the release payload.

Ideally `oc` would also check for the standard
`org.opencontainers.image.{revision,source}` labels too for this purpose
but this assumption is wired in a bunch of places so let's just conform
for now to it.

Add a new option to inject these labels. This can then be used when
building the node image extensions image. In turn, this makes the
release controller page for a release have working links for our images
to the openshift/os commit they were built from, just like the other
images.