kolet: Capture error messages from mkfifo
The Go `exec.Command` when used naively like this captures
stderr and then it gets lost.

It turned out with soft reboot it was this `mkfifo` that was
failing but all we got is "Child process exited with code 1"
but there are like 5 different processes on two different hosts
that could be talking about, and we *really* thought it
was talking about the code-under-test, not the framework setting
it up.