github/dependabot: merge two configs together
Apparently you can't have two config entries for `gomod` so we'll
have to merge these two together. I'll just compromise and go from
daily in one and monthly in the other and settle on weekly for both.

Unfortunately there isn't really a way to validate your dependabot
config before merging the PR and going to the dependabot tab so
we'll YOLO until it works.