cmd-prune: prune blob refs from tmp repo
We now keep blob refs in the tmp repo to make importing OCI images more
efficient by avoiding having to re-import identical layers.

But then we also need to prune them. Do this.

The semantic here is simple: we prune any blob ref not tied to one of
the kept builds.