mantle/aws: rework AWS Windows LI image creation
partially revert and rework d3688be, 0791319, a1f8d97. and 1834b07.

Rework the aws-winli creation logic to use
`aws ec2 register-image --billing-product` to create the AMI with the
Windows License Included billing code metadata. This simplifies the
creation logic and removes the need for swapping the root volume of an
instance. Instead, add a `imageupload-aws` and ore argument
`--winli-billing-product` to set the billing product code during image
creation. Setting billing product codes is limited to approved AWS
accounts, so this will only be used to create the RHCOS aws-winli image.