osbuild: use bootc install to deploy the container
Instead of deploying the container to the tree then copy all the contents
to the disk image, use bootc to directly manage the installation to the
target filesystems.

Right now this requires to use the image as the buildroot so this
requires python (for osbuild). This is tracked in [1].

[1] https://github.com/bootc-dev/bootc/issues/1410
Requires https://github.com/osbuild/osbuild/pull/2149