cosa diff: add support for diffing metal images
This add two flags: `--metal` and `metal-part-table`.

The former will mount the disk partitions using guestfs tools to the
tmp-diff directory, then call git diff over the two mounted filesystems.
It requires multithreading because the FUSE mount call is blocking so
we mount the disks in two libreguest VMs that run in separate threads.

A simpler approach would have been to copy all the content to the
tmp-diff folder but that would copy a lot of data just for diffing. It
may be faster though.

The second flag `--metal-part-table` will simply show the partition
table for the two images.