cmd-build: always insert build before pruning
We had this really subtle interaction going on between cosa build and
prune, where we were actually relying on the prune code to regenerate
`builds.json` to add the new build we just did. There's history there
on how we got here, but at this point it just feels weird. It's cleaner
to always insert the new build ourselves and then separately call
`cosa prune` so let's do that.