cmd-prune: don't prune blob refs from non-existant builds
If the oci manifest doesn't exist locally it's probably because we
buildfetched (which pulls the builds.json for all builds in the stream)
and then did a local build (which happens to be the workflow we use in
the pipeline). Let's just skip it in that case.

Fixes fd9b7da