cmd-build-with-buildah: add FCOS stream label manually
This is for https://github.com/coreos/fedora-coreos-tracker/issues/1996.
I initially wanted to just add another `LABEL` to our Containerfile
and hook that up to another build arg, but the problem is that our
Containerfile is shared with RHCOS/SCOS and there's no way to make
`LABEL` directives conditional.

I opened https://github.com/coreos/rpm-ostree/pull/5454 which will fix
this but for now to not block on this, let's just slap it on from the
cosa side. Once it's folded back into the build process proper, then we
can rever this. But the label not existing shouldn't really affect the
majority of developers anyway.