cmd-build-with-buildah: rework version handling
I think it's still useful to have a `dev` marker in the version string
for development builds.

The versionary script recently learned a new `--dev` switch which will
spit out a development version string. Let's tweak the logic so that:
- by default, if `--versionary` is not passed, we call `versionary
  --dev` to get a development version string
- if `--versionary` is passed, we _don't_ pass `--dev` to get a
  production version string
- if `--version` is passed, it always wins

This is a bit confusing of course. E.g. ideally the `--versionary`
switch would be called `--prod-versionary` or something. But the intent
is to replicate the UX that exists today with `cosa build` so that this
becomes a drop-in replacement for it in the pipeline.

Once we cut over and don't need to support both, we could tweak things.