cmd-diff: support `git difftool` so we can utilize vimdiff
Having vimdiff is a lot more powerful to me because you can manually
do a few things in the terminal to massage the files on each side of the
diff to give you more information (i.e. narrow in on exactly what diff
you are interested in). Let's add a --difftool boolean flag to trigger
`git difftool`, which will allow diffs to be displayed using vimdiff.

Additionally include vim-enhanced so we have `vimdiff` installed and
at our disposal.