cmd-build-with-buildah: in --direct case, tag in containers-storage
Right now, it's an anonymous image, which is awkward. Just add a proper
tag for it so it can more easily be used as is.

I initially tried to just add another `-t` so buildah would output both
a containers-storage tag and an OCI archive, but buildah gets confused
by the first `-t` referring to an OCI archive:

    WARN[0001] don't know how to add tags to images stored in "oci-archive" transport

Will file an issue at some point.

While we're here, also echo the buildah command being run.