testiso: stop checking the ostree commit in the aleph file
This doesn't work in the container-native path because the
ostree-commit in the aleph is from the encapsulation done as part of
`build-chunked-oci`, but the `ostree-commit` we embed in `meta.json` is
from the imported "merge commit".

I _think_ what we need is to tweak our osbuild manifests to use the
import path rather than unencapsulate path, and that should result in
the merge commit getting inserted in the aleph (but also that would be
one less dependency on even having `ostree-commit` in `meta.json` in the
first place).

Anyway, I think just checking the version string is good enough for our
purposes here.