cmd-build-with-buildah: add fedora-coreos.stream as annotation too
In the legacy path, `fedora-coreos.stream` is both an OCI annotation
(which lives in the image manifest) and a label (which lives in the
image config).

Zincati only checks the annotation, not the label. I think checking the
label would've been more appropriate, but anyway for now at least, let's
match the existing semantic in the new container-native path.

Partially fixes https://github.com/coreos/fedora-coreos-tracker/issues/1996