kola/tests/upgrade: tweak for OCI payloads
First, I'm not sure why `:latest` was appended to the filename of OCI
archives. I think that's just a historical artifact. But nowadays, at
least with the new non-encapsulated OCI archives, this causes an error
with skopeo during rebasing:

    error: Preparing import: Fetching manifest: failed to invoke method OpenImage: no descriptor found for reference "latest"

So drop that.

Also, don't rely on `ostree-commit` from `meta.json` yet when
synthesizing another commit to upgrade to, since it's not reproducible
yet. This is fixed in latest bootc, but it still needs to propagate.

(There's a larger picture issue here which is that this test is becoming
less and less realistic as we keep going. We shouldn't be liberally
doing rpm-ostree rebases like this, but instead try to leverage Zincati
more. Also the synthetic update should be fetched as another container
image in a registry, possibly hosted locally, rather than an OSTree
commit.)