platform/metal: don't require metal artifacts for offline scenarios
There's no need for the metal images if we're testing e.g.
`iso-offline-install.bios`. Previously, this didn't matter because we
always had to build the metal and metal4k artifacts to build the live
ISO.

But nowadays with `cosa osbuild`, it's possible to have the live ISO
built without the metal artifacts (well, it's still built, but not
exported out of the osbuild environment). Probably should resolve that,
but dropping that hard requirement on the kola side doesn't hurt either.