set up containers-storage in supermin cache
In the container native build flow where we are building from
quay.io/fedora/fedora-bootc:xx it's better for the local developer
use case if we don't need to pull the container from the registry
on each iteration. Let's use containers-storage from the cache
and also use `buildah build --layers=true` in all cases.

This also bumps the default runvm cache size to account for the
extra usage associated with the containers storage.