cmd-build-with-buildah: unify more the direct and non-direct paths
We now have persistent containers storage in the supermin VM
with the cache. Let's unify the direct and non-direct paths
in the code.

This does mean we need to skopeo copy to the ociarchive inside
the supermin VM (because we don't have access to that containers
storage outside of it). Alternatively we could have tried to
run the `cosa import` inside the supermin VM as well, but the
use of a bare-user repository over a virtiofs share made that
option not work.