cmd-build-with-buildah: add --parent-build for compatibility
The `--parent-build` concept doesn't translate well to the
container-native path. Two main things it was used for was for
maintaining valid OSTree branch history (which is no longer a
concept in the OCI world) and generating good layer chunking
to optimize deltas, which I think the successor to that now is
https://github.com/coreos/rpm-ostree/issues/5460.

For easier ratcheting, allow the switch but make it do nothing.