cmd-build-with-buildah: add --force for compatibility
There is no good change detection right now on the container-native
side. Hermetic Konflux builds should help there, but I'd also like to
make our builds reproducible which is the ultimate solution to this
though less efficient (we'd still do the build, but since we'd get
the same digest, it'd be a no-op -- this is why it's still useful even
with reproducible builds to have good "path change" triggers like what
Konflux offers).