cmd-import: add coreos-assembler.container-config-git to meta.json
That key traditionally represented git information about the source
config repo that was built. But now in the buidah path, the source of
truth for this information has moved to the standard OCI labels. But
for convenience and because a bunch of things check for this in the old
place (notably the build browser and the release job), just proxy that
info under the same key into `meta.json` as before.

This should make the "FCOS commit" key show up in the build browser.