qemu: Ensure virtio journal streaming is killed on shutdown
In https://github.com/ostreedev/ostree/issues/3513 i'm trying to
ensure that all the ostree mounts are cleaned up. It turned
out that it was *this* unit keeping `/var` open because
the `journalctl` binary doesn't know to stop tracking the
`/var/log/journal` files when the daemon has flushed.

(We could obviously fix systemd to handle this)

But anyways the main reason this exists is to forward logs
from startup and especially entering emergency.target. While
it's useful to get logs from shutdown, there are other ways
to do that.

This makes `var.mount` reliably unmounted for me.