cmd-fetch: Derive kofnlus lockfiles from rpm-ostree
Let's use our lockfiles to get hermeto download a set of matching RPMs.
This is done through injecting `includepkg` filters to the repofiles
then feeding that to `rpm-lockfile-prototype`.

The overview is:
we parse the manifest for a package list and enabled repos.
For each package, we take the expected EVRA in the rpm-ostree lockfile
and inject it in the `includepkg` filter in the repos files.
Finally we feed that package list and the custom repo definition to
rpm-ostree-prototype.
The result is a hermeto/cachi2 lockfile we will commit to the repo.

This `--konflux` flag would then be added to the Jenkins bump-lockfiles
job to update both lockfiles in the same commit.

This will allow hermetic builds in Konflux.

I also added a sanity check at the end that iterate over both lockfiles
and compare all versions string to make sure we have exact matches all
around.

See also https://github.com/coreos/fedora-coreos-config/pull/3644 for
the inital implementation.