cmd-sign: add support for signing OCI images
This adds a new `cosa sign --oci` command to sign OCI container images.
This is part of the effort to move FCOS to a container-native build
flow, where we now produce non-encapsulated container images.

The new command works by sending a request to Robosignatory to sign
the image manifest digest. Robosignatory returns a detached signature,
which we then merge with the original payload to create a cleartext
signed message that can be understood by containers/image.

This is a short-term solution until we can move to Sigstore.

Part of https://github.com/coreos/fedora-coreos-tracker/issues/1969.