cosalib/aws: don't require build artifact for AWS WinLI builds
The build artifact is not needed in the AWS WIndows License Included
(WinLI) case because the image is built from an existing snapshot.
To allow individual cloud providers to override this requirement,
move the build artifact check to each provider's wrapper. Don't
perform the check in the AWS wrapper if building a winli image.

Also modify the AWS ore upload code to not require an upload file
when a source snapshot is used, which allows the WinLI image to be
built without the need for an upload file.