cmd-sign: make staging location stream-specific
We want to store all the signatures in the same location rather than
stream-specific. But ideally we still want the staging location for
signing to be stream-specific so that we can safely garbage collect
stale files there without worrying about stepping on concurrent runs.

Just pick up the stream from the metadata and use that to build the
staging location.

See also https://github.com/coreos/fedora-coreos-pipeline/pull/1218.