release: don't ostree commits into release metadata on f43+
Starting in f43, we won't be publishing content to the OSTree repo
anymore, so there's no point in having any commit hash in the release
metadata.

On builds that have `oci-imported: true`, let's stop adding the `commit`
key in the release metadata in `cosa generate-release-metadata`. That'll
naturally bind it to f43.

In plume, handle the case where there is no commit string.