release: stop putting ostree commits into release metadata
We've fully cut over to OCI now for updates so we can stop having ostree
commit info in there. It'll make even less sense with the buildah path
where the commit hash is synthetic.

In `cosa generate-release-metadata`, stop propagating the commit. In
plume, stop propagating the commit to the release index.