plume/release: drop metadata URL check
When we update the release index, plume checks if the release we want to
add already exists by checking for a release with the same version _and_
metadata URL. The latter check seems unnecessary and in fact harmful
because there shouldn't ever be two releases in the list with the same
version number at all.

So drop that check.

Noticed this when testing that code locally (with `--local-mode`) in
which we don't pass S3 information, so it's normal that the metadata URL
isn't the same.