cmd-diff: fix rpm diffing in a build-with-buildah world
In the build-with-buildah world there is no ostree-commit-object
so there is no lightweight (i.e. partial import) way to do a db diff.
We have to do a full import here unless we somehow start storing RPM
NEVRAs in a label or annotation.

Also, we need to pass the version (id) to `db diff` because the
commit checksum in the metadata doesn't seem to work when diffing
a build-with-buildah image.