cosalib/aws: only replicate WinLI AMIs when --winli is specified
The current logic always considers the `aws-winli` buildmeta key when
present, which causes WinLI AMIs to be replicated in both the standard
AWS and GovCloud cases.

Restore `aws_run_ore_replicate`[1] to only work on the 'amis' buildmeta
key and switch to aws-winli when `--winli` is used. This ensures WinLI
replication happens only when explicitly requested. Also update the
CLI help text for `--winli` to clarify that it applies to replication
as well as creation.

[1]: https://github.com/coreos/coreos-assembler/pull/4069/commits/a5bfbdefe999d5cf4ec5b6d4c1de28798b6322d6