move diff calculations into `cosa import`
This moves the RPM/Advisory diffing into `cosa import`, which
means we'll get that info even when importing from another source.

It also changes the code such that we'll only attempt to add that
info if there was something to diff against. i.e. for a first
build in a stream (like when doing a local dev build for the first
time) now we won't error out.

Fixes https://github.com/coreos/coreos-assembler/issues/4336