osbuild/ bootc-install: symlink aleph files
Create symlinks to the aleph file created by bootc so our tests and
tooling find the aleph at the expected path.

Note that when moving to image-build we will likely move this to
an overlay in the config, that's way easier than having to wire up a
blueprint option to allow creating arbitrary symlinks.