vmdeps: add filesystem to supermin
this rpm provides locales under `/usr/share/locale`.
grub2 install these files inside the /boot partition using what
exist on the host. Mooving to bootc install results in an empty
folder because those files don't exist in the buildroot, so add them.

In the previous manifest we were using a chroot before calling bootupd
which made sure bootupd (and by extension grub2) would pull content
from the target root.

It would be nicer to have bootc install also chroot before calling
bootupd, it would have the added benefit to use the bootupd (and other)
binaries that are shipped on the image rather than what is in the buildroot.

Again, this is a complication coming from not having python in the image
so we can't use ourselves as the buildroot, which would alleviate a lot
of pain.

See https://github.com/coreos/coreos-assembler/pull/4224#issuecomment-3452307787
And https://github.com/bootc-dev/bootc/issues/1559