osbuild: use bootc install to deploy the container
Instead of deploying the container to the tree then copy all the contents
to the disk image, use bootc to directly manage the installation to the
target filesystems.

Right now this requires to use the image as the buildroot so this
requires python (for osbuild). This is tracked in [1].
As we have python in rawhide now I duplicated the manifest and added a
switch in the osbuild wrapper script.

We can keep the manifest duplicated until we are confident to roll this
to all streams.

[1] https://github.com/bootc-dev/bootc/issues/1410

Requires:
https://github.com/bootc-dev/bootc/pull/1460
https://github.com/bootc-dev/bootc/pull/1451
https://github.com/osbuild/osbuild/pull/2149
https://github.com/osbuild/osbuild/pull/2152

All of which have landed in osbuild-159 and bootc 1.6