cmdlib: add excludepkgs for repos from fast-tracks.yaml
For some reason I was seeing behavior where more packages were getting
pulled from a repo generated with fast-tracks.yaml [1] than just what
was in the includepkgs line. I found that adding `excludepkgs=*` worked
but according to the docs that shouldn't be needed.

Life is too short, so let's just add it here.

[1] https://github.com/coreos/coreos-assembler/pull/4058