osbuild: set agcount=2 for XFS filesystems
In the past the agcount has been 4 (I think the default),
but I've been told that it may also depend on your the
number of processors in the host where you run mkfs.xfs, so
it could be higher than that.

We want to set it to the lowest possible value here (2) so
growing filesystems won't cause our agcount to increase as
dramatically as it has in the past.

Closes https://github.com/coreos/fedora-coreos-tracker/issues/1993