src/konflux-rpm-lockfile: use coreos-pool URLs only
Until now, we tackled the DNF5 issue [1] by prioritizing the URL
location of the repo 'coreos-pool' when multiple repo were found
for the say locked NEVRA.
But, if that locked NEVRA has only one URL location coming from
the fedora repos i.e: fedora, fedora-updates, then the repoquery
command returns a broken URL location for ppc64le and s390x arches
because of the DNF5 issue. It returns the URL from the first item
defined in baseurl, whereas the resolution worked with the second
item i.e fedora-secondary repo where the RPMs for ppc64le and
s390x are available.

This patch fixes this issue with another approach wich consists
of using coreos-pool URLs only. For each non coreos-pool URL we
extract the NEVRA and arch to construct the URL into the
canonical koji path structure. As soon as the 'bump-lockfile'
commit is merged, the coreos-koji-tagger will tag to 'coreos-pool'.
So it's safe to craft the URL prior its availability.
One downside is a possible race condition between the koji
tag operation and the CI/CD system building hermetically. The other
downside is that the `rpms.lock.yaml` changes cannot be tested
locally as this relies on remote koji tagging operation. But,
as it's consumed for Konflux only now, I'd say it's ok.

[1] https://github.com/rpm-software-management/dnf5/issues/2466