src/cmd-build-with-buildah: add konflux lockfile option
We add the option to the container native build command that enables
generating the konflux lockfile at the end of the build.
We could implicitly gererate the konflux lockfiles alongside the
classic ones, but this has a high cost of fetching the repo
metadata in order to repoquery for each NEVRA. This operation is
only required to bump the lockfiles in Jenkins, so it does not
make sense to pay the cost for every build.